auth_services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.6.0
python-jose==3.3.0
email-validator==2.3.0
httpx==0.25.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
